+++ src/algos/avoidChange.ts
@@ -46,18 +46,18 @@ export function avoidChange({
   validateFeeRate(feeRate);
   validateFeeRate(dustRelayFeeRate);
 
-  const targetsValue = targets.reduce((a, target) => a + target.value, 0);
+  const targetsValue = targets.reduce((a, target) => a + target.value, 0n);
   const utxosSoFar: Array<OutputWithValue> = [];
 
   for (const candidate of utxos) {
-    const utxosSoFarValue = utxosSoFar.reduce((a, utxo) => a + utxo.value, 0);
+    const utxosSoFarValue = utxosSoFar.reduce((a, utxo) => a + utxo.value, 0n);
 
     const txSizeWithCandidateAndChange = vsize(
       [candidate.output, ...utxosSoFar.map(utxo => utxo.output)],
       [remainder, ...targets.map(target => target.output)]
     );
-    const txFeeWithCandidateAndChange = Math.ceil(
-      txSizeWithCandidateAndChange * feeRate
+    const txFeeWithCandidateAndChange = BigInt(
+      Math.ceil(txSizeWithCandidateAndChange * feeRate)
     );
 
     const remainderValue =
@@ -69,15 +69,15 @@ export function avoidChange({
       [candidate.output, ...utxosSoFar.map(utxo => utxo.output)],
       targets.map(target => target.output)
     );
-    const txFeeWithCandidate = Math.ceil(txSizeWithCandidate * feeRate);
+    const txFeeWithCandidate = BigInt(Math.ceil(txSizeWithCandidate * feeRate));
 
     const txSizeSoFar = vsize(
       utxosSoFar.map(utxo => utxo.output),
       targets.map(target => target.output)
     );
-    const txFeeSoFar = Math.ceil(txSizeSoFar * feeRate);
+    const txFeeSoFar = BigInt(Math.ceil(txSizeSoFar * feeRate));
     const candidateFeeContribution = txFeeWithCandidate - txFeeSoFar;
-    if (candidateFeeContribution < 0)
+    if (candidateFeeContribution < 0n)
       throw new Error(`candidateFeeContribution < 0`);
 
     // Only consider inputs with more value than the fee they require
